Fontello icon names and GitHub URLs
Sometimes, a Fontello glyph name have a numbered suffix. This suffix can be the
name of the pictogram (e.g. "progress-5"), or it could have been added
because another pictogram (in another collection) have the same name
(e.g.  "search-2"). In the second case, the suffix need to be removed.

This commit also adds the remaining GitHub URL exceptions.